added validation on procstat values
